Move isWebdriver into public-lib so as not to break public-lib
