+++ src/public-lib/lib/webdriverUtils.ts
@@ -1,0 +1,3 @@
+export function isWebdriver() {
+    return window.navigator.webdriver;
+}
